+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 ## (UNRELEASED)
 
 ### FEATURES:
+* builder/docker: add packer docker builder support
 * builder/file: add packer file builder support
 * provisioner/file: add packer file provisioner support